requirements.txt
langchain==0.1.20
langchain-core==0.1.52
langchain-groq==0.1.0
pydantic==2.10.0
groq==0.16.0
python-dotenv==1.0.0
numpy==1.26.4
requests>=2.31.0